Implemented error handling and made some UI improvements

+++ Sideloader/Services/PackageManager.cs
@@ -13,24 +13,63 @@ namespace Sideloader.Services
     {
         public async Task RetrievePackage(AppPackage appPackage)
         {
-            var downloadUrl = appPackage.DownloadUrl;
+            Cleanup("package" + appPackage.Platform.Name);
+            OnPackageStatusChanged("Downloading package...");
+            var directoryName = await DownloadFile(appPackage);
+            OnPackageStatusChanged("Extracting zip content...");
+            ExtractZip(directoryName);
+            appPackage.InstallFolderPath = GetFolderPath(directoryName);
+            OnPackageStatusChanged("Installing dependencies...");
+            InstallDependencies(appPackage);
+            OnPackageStatusChanged("Installing package...");
+            InstallPackage(appPackage);
+        }
+
+        private void ExtractZip(string directoryName)
+        {
+            try
+            {
+                ZipFile.ExtractToDirectory("package.zip", directoryName);
+            }
+            catch (Exception exception)
+            {
+                OnPackageStatusChanged("Unzip failed. " + exception.Message, MessageType.Error);
+                throw new PackageException(exception.ToString(), true);
+            }
+        }
+
+        private async Task<string> DownloadFile(AppPackage appPackage)
+        {
             var webClient = new System.Net.WebClient();
             var directoryName = "package" + appPackage.Platform.Name;
             try
             {
-                File.Delete("package.zip");
-                Directory.Delete(directoryName, true);
+                await webClient.DownloadFileTaskAsync(new Uri(appPackage.DownloadUrl), "package.zip");
             }
             catch (Exception exception)
             {
-                Logger.Instance.Error(exception);
+                OnPackageStatusChanged("Package download failed. DevInfo: " + exception.Message, MessageType.Error);
+                throw new PackageException(exception.ToString(), true);
             }
-            await webClient.DownloadFileTaskAsync(new Uri(downloadUrl), "package.zip");
-            ZipFile.ExtractToDirectory("package.zip", directoryName);
-            appPackage.InstallFolderPath = GetFolderPath(directoryName);
-            InstallDependencies(appPackage);
-            InstallPackage(appPackage);
+            return directoryName;
         }
+
+        private void Cleanup(string directoryName)
+        {
+            try
+            {
+                File.Delete("package.zip");
+                if (Directory.Exists(directoryName))
+                    Directory.Delete(directoryName, true);
+            }
+            catch (Exception exception)
+            {
+                OnPackageStatusChanged("Cleanup failed. " + exception.Message, MessageType.Error);
+                throw new PackageException(exception.ToString(), true);
+            }
+        }
+
+        public event EventHandler<StatusInfo> PackageStatusChanged;
 
         private string GetFolderPath(string path)
         {
@@ -60,13 +99,28 @@ namespace Sideloader.Services
             string status = process.StandardOutput.ReadToEnd();
             Debug.WriteLine(status);
             process.WaitForExit();
+            if (string.IsNullOrWhiteSpace(status) == false)
+            {
+                OnPackageStatusChanged("Can't install dependency. " + status, MessageType.Error);
+                throw new PackageException("Can't install dependency. " + status, true);
+            }
         }
 
         private string GetDependency(AppPackage package)
         {
-            var path = package.InstallFolderPath + @"\\Dependencies\\" + package.Platform.Name;
-            var fileNames = Directory.GetFiles(path);
-            return Path.GetFileName(fileNames[0]);
+            try
+            {
+                var path = package.InstallFolderPath + @"\\Dependencies\\" + package.Platform.Name;
+                var fileNames = Directory.GetFiles(path);
+                var dependencyPath = ".\\packagex86\\" + Path.Combine(path, Path.GetFileName(fileNames[0])).Split(new[] { "\\packagex86" }, StringSplitOptions.None)[1];
+                return dependencyPath;
+            }
+            catch (Exception exception)
+            {
+                OnPackageStatusChanged("Can't resolve dependency. " + exception.Message, MessageType.Error);
+                Logger.Instance.Error("Dependency. " + exception);
+                throw new PackageException("Can't resolve dependency. " + exception.Message, true);
+            }
         }
 
         private void InstallPackage(AppPackage package)
@@ -78,27 +132,69 @@ namespace Sideloader.Services
                 {
                     File.GetAttributes(path);
                     string strCmdText = Path.Combine(package.InstallFolderPath, "Add-AppDevPackage.ps1");
-                    var process = new Process();
-                    process.StartInfo.UseShellExecute = false;
-                    process.StartInfo.RedirectStandardOutput = true;
-                    process.StartInfo.FileName = @"C:\windows\system32\windowspowershell\v1.0\powershell.exe";
-                    process.StartInfo.Arguments = "\"&'" + strCmdText + "'\"";
-                    process.StartInfo.WorkingDirectory = package.InstallFolderPath;
+                    var process = new Process
+                    {
+                        StartInfo =
+                        {
+                            UseShellExecute = false,
+                            RedirectStandardOutput = true,
+                            FileName = @"C:\windows\system32\windowspowershell\v1.0\powershell.exe",
+                            Arguments = "\"&'" + strCmdText + "'\"",
+                            WorkingDirectory = package.InstallFolderPath
+                        }
+                    };
 
                     process.Start();
-                    string s = process.StandardOutput.ReadToEnd();
+                    string installInfo = process.StandardOutput.ReadToEnd();
+                    Logger.Instance.Info(installInfo);
                     process.WaitForExit();
-                    using (StreamWriter outfile = new StreamWriter("StandardOutput.txt", true))
+                    if (installInfo.Contains("Error: "))
                     {
-                        outfile.Write(s);
+                        var message = installInfo;
+                        if (installInfo.Contains("The current user has already installed"))
+                        {
+                            message = "Please uninstall the application first";
+                        }
+                        OnPackageStatusChanged(message, MessageType.Error);
+                    }
+                    using (StreamWriter outfile = new StreamWriter("install_log.txt", true))
+                    {
+                        outfile.Write(installInfo);
                     }
 
                 }
             }
             catch (Exception exception)
             {
-                Debug.WriteLine(exception.Message);
+                OnPackageStatusChanged("Package installation failed. " + exception.Message, MessageType.Error);
+                Logger.Instance.Error("Package instllation: ", exception);
+                throw new PackageException("Package installation failed", true);
             }
         }
+
+        protected virtual void OnPackageStatusChanged(string message, MessageType messageType = MessageType.Info)
+        {
+            Logger.Instance.Info("STATUS: " + message);
+            var messageInfo = new StatusInfo(messageType, message);
+            PackageStatusChanged?.Invoke(this, messageInfo);
+        }
+    }
+
+    public class StatusInfo
+    {
+        public MessageType MessageType { get; }
+        public string Message { get; }
+
+        public StatusInfo(MessageType messageType, string message)
+        {
+            MessageType = messageType;
+            Message = message;
+        }
+    }
+
+    public enum MessageType
+    {
+        Info,
+        Error
     }
 }
